@@ -1,9 +1,9 @@
 module github.com/knqyf263/pet
 
-go 1.17
+go 1.20
 
 require (
-	github.com/BurntSushi/toml v0.3.0
+	github.com/atotto/clipboard v0.1.4
 	github.com/briandowns/spinner v0.0.0-20170614154858-48dbb65d7bd5
 	github.com/chzyer/logex v1.1.10 // indirect
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
@@ -11,14 +11,15 @@ require (
 	github.com/fatih/color v1.7.0
 	github.com/google/go-github v15.0.0+incompatible
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/jroimartin/gocui v0.4.0
-	github.com/mattn/go-runewidth v0.0.2
+	github.com/mattn/go-colorable v0.0.9 // indirect
+	github.com/mattn/go-isatty v0.0.3 // indirect
+	github.com/mattn/go-runewidth v0.0.10
 	github.com/pkg/errors v0.8.0
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.1 // indirect
 	github.com/xanzy/go-gitlab v0.50.3
 	//github.com/xanzy/go-gitlab v0.10.5
-	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
+	golang.org/x/crypto v0.17.0
 	golang.org/x/oauth2 v0.0.0-20181106182150-f42d05182288
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
 )
@@ -29,16 +30,25 @@ require (
 )
 
 require (
+	github.com/awesome-gocui/gocui v1.1.0
+	github.com/go-test/deep v1.1.0
+	github.com/pelletier/go-toml v1.9.5
+)
+
+require (
 	github.com/alessio/shellescape v1.4.1 // indirect
+	github.com/gdamore/encoding v1.0.0 // indirect
+	github.com/gdamore/tcell/v2 v2.4.0 // indirect
 	github.com/golang/protobuf v1.2.0 // indirect
 	github.com/google/go-querystring v1.0.0 // indirect
 	github.com/hashicorp/go-cleanhttp v0.5.1 // indirect
 	github.com/hashicorp/go-retryablehttp v0.6.8 // indirect
-	github.com/mattn/go-colorable v0.0.9 // indirect
-	github.com/mattn/go-isatty v0.0.3 // indirect
-	github.com/nsf/termbox-go v0.0.0-20180509163535-21a4d435a862 // indirect
-	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
-	golang.org/x/sys v0.0.0-20211116061358-0a5406a5449c // indirect
+	github.com/lucasb-eyer/go-colorful v1.0.3 // indirect
+	github.com/rivo/uniseg v0.1.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/term v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
 	golang.org/x/time v0.0.0-20191024005414-555d28b269f0 // indirect
 	google.golang.org/appengine v1.3.0 // indirect
 )
