@@ -1,8 +1,6 @@
 package cmd
 
-import (
-	"fmt"
-	"io/ioutil"
+import "os"
 
 	"github.com/knqyf263/pet/config"
 	petSync "github.com/knqyf263/pet/sync"
@@ -43,7 +41,7 @@ func edit(cmd *cobra.Command, args []string) (err error) {
 	// file content before editing
 	before := fileContent(snippetFile)
 
-	err = editFile(editor, snippetFile)
+	err = editFile(editor, snippetFile, 0)
 	if err != nil {
 		return
 	}
@@ -64,7 +62,7 @@ func edit(cmd *cobra.Command, args []string) (err error) {
 }
 
 func fileContent(fname string) string {
-	data, _ := ioutil.ReadFile(fname)
+	data, _ := os.ReadFile(fname)
 	return string(data)
 }
 
