@@ -14,43 +14,55 @@ import (
 )
 
 const (
-	githubTokenEnvVariable = "PET_GITHUB_ACCESS_TOKEN"
+	gheTokenEnvVariable = "PET_GHE_ACCESS_TOKEN"
 )
 
-// GistClient manages communication with Gist
-type GistClient struct {
+// GHEGistClient manages communication with Gist
+type GHEGistClient struct {
 	Client *github.Client
 	ID     string
 }
 
-// NewGistClient returns GistClient
-func NewGistClient() (Client, error) {
-	accessToken, err := getGithubAccessToken()
+// NewGHEGistClient returns GistClient
+func NewGHEGistClient() (Client, error) {
+	accessToken, err := getGHEAccessToken()
 	if err != nil {
 		return nil, fmt.Errorf(`access_token is empty.
 Go https://github.com/settings/tokens/new and create access_token (only need "gist" scope).
 Write access_token in config file (pet configure) or export $%v.
-		`, githubTokenEnvVariable)
+		`, gheTokenEnvVariable)
 	}
 
-	client := GistClient{
-		Client: githubClient(accessToken),
-		ID:     config.Conf.Gist.GistID,
+	var baseUrl, uploadUrl string
+
+	if config.Conf.GHEGist.BaseUrl != "" {
+		fmt.Println(config.Conf.GHEGist.BaseUrl)
+		baseUrl = config.Conf.GHEGist.BaseUrl
+	}
+
+	if config.Conf.GHEGist.UploadUrl != "" {
+		fmt.Println(config.Conf.GHEGist.UploadUrl)
+		uploadUrl = config.Conf.GHEGist.UploadUrl
+	}
+
+	client := GHEGistClient{
+		Client: githubEnterpriseClient(accessToken, baseUrl, uploadUrl),
+		ID:     config.Conf.GHEGist.GistID,
 	}
 	return client, nil
 }
 
-func getGithubAccessToken() (string, error) {
-	if config.Conf.Gist.AccessToken != "" {
-		return config.Conf.Gist.AccessToken, nil
-	} else if os.Getenv(githubTokenEnvVariable) != "" {
-		return os.Getenv(githubTokenEnvVariable), nil
+func getGHEAccessToken() (string, error) {
+	if config.Conf.GHEGist.AccessToken != "" {
+		return config.Conf.GHEGist.AccessToken, nil
+	} else if os.Getenv(gheTokenEnvVariable) != "" {
+		return os.Getenv(gheTokenEnvVariable), nil
 	}
-	return "", errors.New("Github AccessToken not found in any source")
+	return "", errors.New("GHE AccessToken not found in any source")
 }
 
 // GetSnippet returns the remote snippet
-func (g GistClient) GetSnippet() (*Snippet, error) {
+func (g GHEGistClient) GetSnippet() (*Snippet, error) {
 	s := spinner.New(spinner.CharSets[14], 100*time.Millisecond)
 	s.Start()
 	s.Suffix = " Getting Gist..."
@@ -69,7 +81,7 @@ func (g GistClient) GetSnippet() (*Snippet, error) {
 	}
 
 	content := ""
-	filename := config.Conf.Gist.FileName
+	filename := config.Conf.GHEGist.FileName
 	for _, file := range gist.Files {
 		if *file.Filename == filename {
 			content = *file.Content
@@ -86,12 +98,12 @@ func (g GistClient) GetSnippet() (*Snippet, error) {
 }
 
 // UploadSnippet uploads local snippets to Gist
-func (g GistClient) UploadSnippet(content string) error {
+func (g GHEGistClient) UploadSnippet(content string) error {
 	gist := &github.Gist{
 		Description: github.String("description"),
-		Public:      github.Bool(config.Conf.Gist.Public),
+		Public:      github.Bool(config.Conf.GHEGist.Public),
 		Files: map[github.GistFilename]github.GistFile{
-			github.GistFilename(config.Conf.Gist.FileName): github.GistFile{
+			github.GistFilename(config.Conf.GHEGist.FileName): github.GistFile{
 				Content: github.String(content),
 			},
 		},
@@ -111,7 +123,7 @@ func (g GistClient) UploadSnippet(content string) error {
 	return nil
 }
 
-func (g GistClient) createGist(ctx context.Context, gist *github.Gist) (gistID *string, err error) {
+func (g GHEGistClient) createGist(ctx context.Context, gist *github.Gist) (gistID *string, err error) {
 	s := spinner.New(spinner.CharSets[14], 100*time.Millisecond)
 	s.Start()
 	s.Suffix = " Creating Gist..."
@@ -124,7 +136,7 @@ func (g GistClient) createGist(ctx context.Context, gist *github.Gist) (gistID *
 	return retGist.ID, nil
 }
 
-func (g GistClient) updateGist(ctx context.Context, gist *github.Gist) (err error) {
+func (g GHEGistClient) updateGist(ctx context.Context, gist *github.Gist) (err error) {
 	s := spinner.New(spinner.CharSets[14], 100*time.Millisecond)
 	s.Start()
 	s.Suffix = " Updating Gist..."
@@ -136,11 +148,11 @@ func (g GistClient) updateGist(ctx context.Context, gist *github.Gist) (err erro
 	return nil
 }
 
-func githubClient(accessToken string) *github.Client {
+func githubEnterpriseClient(accessToken, baseURL, uploadURL string) *github.Client {
 	ts := oauth2.StaticTokenSource(
 		&oauth2.Token{AccessToken: accessToken},
 	)
-	tc := oauth2.NewClient(oauth2.NoContext, ts)
-	client := github.NewClient(tc)
+	tc := oauth2.NewClient(context.Background(), ts)
+	client, _ := github.NewEnterpriseClient(baseURL, uploadURL, tc)
 	return client
 }
